# implementation/src/main/java/com/siller/rohlik/store/order/model/CreateNewOrderError.java
package com.siller.rohlik.store.order.model;

import lombok.AllArgsConstructor;
import lombok.Getter;

@AllArgsConstructor
@Getter
public class CreateNewOrderError {
    private String productId;
    private Code errorCode;

    public enum Code {
        INVALID_PRODUCT,
        NOT_ENOUGH_PRODUCTS_ON_STOCK
    }

}